+++ CHANGELOG.rst
@@ -10,6 +10,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on KeepAChangelog_ and this project adheres to SemanticVersioning_.
 
+[Unreleased]
+++++++++++++
+
+Changed
+~~~~~~~
+* Updates to django-webix 1.4.0 version
+
 
 [0.1.0] - 2018-09-27
 ++++++++++++++++++++